pc_agent/requirements.txt
websockets>=12.0
pyautogui>=0.9.54
pyperclip>=1.8
pywin32>=306
Pillow>=10.0
psutil>=5.9
pyinstaller>=6.0
